requirements.txt
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
numpy==2.3.3
pandas==2.3.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.23
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
pyyaml
rapidfuzz